# [Scala Native] Use alt custom priority queue