Main: `Scala Native | 305.46 ms | 3.85 s | 31.87 s`

This PR:
```
Scala Native:

        Benchmark 1: ./target/scala-3.3.1/scala_native-out
        Processing time (w/o IO): 257ms
        Processing time (w/o IO): 265ms
        Processing time (w/o IO): 265ms
        Processing time (w/o IO): 257ms
        Processing time (w/o IO): 256ms
        Processing time (w/o IO): 257ms
        Processing time (w/o IO): 257ms
        Processing time (w/o IO): 258ms
        Processing time (w/o IO): 268ms
        Processing time (w/o IO): 273ms
        Processing time (w/o IO): 258ms
        Processing time (w/o IO): 275ms
        Processing time (w/o IO): 275ms
          Time (mean ± σ):     344.6 ms ±   9.3 ms    [User: 323.1 ms, System: 79.4 ms]
          Range (min … max):   336.1 ms … 357.8 ms    10 runs
         
Scala Native:

        Benchmark 1: ./target/scala-3.3.1/scala_native-out
        Processing time (w/o IO): 3713ms
        Processing time (w/o IO): 3667ms
          Time (abs ≡):         4.043 s               [User: 4.366 s, System: 0.419 s]
         
Scala Native:

        Benchmark 1: ./target/scala-3.3.1/scala_native-out
        Processing time (w/o IO): 31681ms
        Processing time (w/o IO): 31128ms
          Time (abs ≡):        32.377 s               [User: 37.090 s, System: 1.275 s]
```