@jinyus 2023-11-14
Fresh run from main:
```
Scala Native:

        Benchmark 1: ./target/scala-3.3.1/scala_native-out
        Processing time (w/o IO): 244ms
        Processing time (w/o IO): 243ms
        Processing time (w/o IO): 251ms
        Processing time (w/o IO): 244ms
        Processing time (w/o IO): 263ms
        Processing time (w/o IO): 243ms
        Processing time (w/o IO): 259ms
        Processing time (w/o IO): 242ms
        Processing time (w/o IO): 261ms
        Processing time (w/o IO): 243ms
        Processing time (w/o IO): 243ms
        Processing time (w/o IO): 242ms
        Processing time (w/o IO): 245ms
          Time (mean ± σ):     329.0 ms ±  10.3 ms    [User: 307.3 ms, System: 77.6 ms]
          Range (min … max):   321.1 ms … 348.3 ms    10 runs
         
Scala Native:

        Benchmark 1: ./target/scala-3.3.1/scala_native-out
        Processing time (w/o IO): 3356ms
        Processing time (w/o IO): 3341ms
          Time (abs ≡):         3.671 s               [User: 4.016 s, System: 0.352 s]
         
Scala Native:

        Benchmark 1: ./target/scala-3.3.1/scala_native-out
        Processing time (w/o IO): 28909ms
        Processing time (w/o IO): 28866ms
          Time (abs ≡):        30.088 s               [User: 34.606 s, System: 1.174 s]
```